Squashed commit of the following:
commit 9d36ee9c9c816d7ef55670ddfae52d947a7e1ae0
Author: Locxion <locxion.m@googlemail.com>
Date:   Sun Feb 19 21:58:10 2023 +0100

    Added Helper Class
    Added Summaries to Methods
